requirements.txt
fastapi
uvicorn
pydantic
email-validator
bcrypt
pymongo
python-multipart
torch
torchvision
timm
pillow
python-dotenv
requests
dnspython